requirements.txt
django
dj-stripe==2.8.4
# stripe>=7.0.0,<8.0.0
openai
anthropic
mistralai
daphne
channels
redis
tweepy
celery
djangorestframework
channels_redis
hubspot-api-client
huggingface_hub
gTTS
vosk
shortuuid
facebook-sdk
facebook-business
beautifulsoup4
selenium
django-celery-beat
social-auth-app-django